further refactoring of parking code.

+++ src/main/java/ch/ethz/idsc/amodeus/parking/capacities/ParkingCapacityLinkLength.java
@@ -7,7 +7,7 @@ public class ParkingCapacityLinkLength extends ParkingCapacityAbstract {
 
     /** Assigns to every link in the @param network a capacity equal to:
      * min(@param minCapacityGlobal, link lenghth x @param capacityPerLengthUnit ) */
-    public ParkingCapacityLinkLength(Network network, double capacityPerLengthUnit, double minCapacityGlobal) {
+    public ParkingCapacityLinkLength(Network network, double capacityPerLengthUnit, long minCapacityGlobal) {
         for (Link link : network.getLinks().values()) {
             long capacity = (long) Math.max(Math.floor(link.getLength() * capacityPerLengthUnit), minCapacityGlobal);
             capacities.put(link.getId(), capacity);
